% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_timing.R
\name{add_manufacturer}
\alias{add_manufacturer}
\title{Add field for manufacturers}
\usage{
add_manufacturer(df)
}
\arguments{
\item{df}{a laps df}
}
\value{
a tibble
}
\description{
Add field for manufacturers
}
