% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_tables.R
\docType{data}
\name{driver_teams}
\alias{driver_teams}
\title{List of drivers and teams}
\format{
An object of class \code{character} of length 28.
}
\usage{
driver_teams
}
\description{
Driver last name and team name
}
\keyword{datasets}
