% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_timing.R
\name{filter_slow}
\alias{filter_slow}
\title{Filter Out Slow Laps}
\usage{
filter_slow(df, race_pace = 1.07)
}
\arguments{
\item{df}{a laps df}

\item{race_pace}{(double): percentage over fastest lap to filter out. Defaults to 1.07 (107\%).}
}
\value{
a tibble
}
\description{
Filters laps slower than designated race pace (defaults to 107\% of fastest
lap time).
}
\examples{
fast_laps <- laps \%>\% preprocess_laps() \%>\%  filter_slow(race_pace=1.1)
}
