% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_timing.R
\name{filter_yellow}
\alias{filter_yellow}
\title{Filter Out Yellow Flag Laps}
\usage{
filter_yellow(df, keep_yellows = FALSE)
}
\arguments{
\item{df}{a laps df}

\item{keep_yellows}{(bool): whether to return only the yellow laps. Defaults to False.}
}
\value{
tibble
}
\description{
Filters out yellow flag laps.
}
\examples{
no_yellows <- laps \%>\% preprocess_laps() \%>\% filter_yellow()
}
