% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/racetools.R
\name{get_data}
\alias{get_data}
\title{Get a dataframe of session data from racetools.com}
\usage{
get_data(
  event,
  session,
  root_url = "https://racetools.com/logfiles/IndyCar/2023/"
)
}
\arguments{
\item{event}{(string): the name of the event weekend (ex. "Toronto").}

\item{session}{(string): c("Practice", "Qualifications", "Race")}

\item{root_url}{(string): the url to start the search in. You would only
update this if you wanted a different year's data.}
}
\value{
dataframe
}
\description{
Provide the event and session and we will attempt to find the zip file,
download and unzip it, and return the dataframe.
}
