% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_timing.R
\name{get_fastest}
\alias{get_fastest}
\title{Get the fastest lap for a list of drivers}
\usage{
get_fastest(df, drivers = unique(df$driver_name))
}
\arguments{
\item{df}{A dataframe of laptimes and drivers}

\item{drivers}{The list of drivers to return. Defaults to all.}
}
\value{
a dataframe
}
\description{
Get the fastest lap for a list of drivers
}
\examples{
get_fastest(df = laps, drivers = c("Grosjean", "O'Ward"))
}
