% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_tables.R
\name{get_team_colors}
\alias{get_team_colors}
\title{Get Team Colors For a Given Team}
\usage{
get_team_colors(team_name)
}
\arguments{
\item{team_name}{(str): team name}
}
\value{
(character): hex color of team
}
\description{
Gets a team's primary colors from a team name
}
