% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/racetools.R
\name{get_url}
\alias{get_url}
\title{Get the URL to a session zip file from racetools.com.}
\usage{
get_url(root_url = "https://racetools.com/logfiles/IndyCar/2023/")
}
\arguments{
\item{root_url}{the base url to do the search for. Defaults to the current season in Indycar on racetools.com}
}
\value{
str: the url of the session of interest
}
\description{
This function will prompt you several times for which race you are interested in.
}
