% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_themes.R
\name{indycar_colors}
\alias{indycar_colors}
\title{Get a list of indycar colors and hexes}
\usage{
indycar_colors(...)
}
\arguments{
\item{...}{either NULL or a character vector of colors in the palette you wish
to return hex values for.}
}
\description{
Simply call indycar_colors() to get a list of our
theme colors and their hex values, or provide a color name and it will
return the hex.
}
