% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_themes.R
\docType{data}
\name{indycar_theme}
\alias{indycar_theme}
\title{Theme for prettifying ggplots}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 11.
}
\usage{
indycar_theme
}
\description{
Theme. Simply use + indycar_theme to add this theme to any custom ggplot2 object
}
\keyword{datasets}
