% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_timing.R
\name{load_data}
\alias{load_data}
\title{Load Data from racetools.com}
\usage{
load_data(url)
}
\arguments{
\item{url}{A url to a zip file (ideally from racetools.com)}
}
\value{
a tibble of laps data
}
\description{
Loads a zip file from the internet, unzips it, and returns it as a tibble dataframe.
}
