% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/racetools.R
\name{save_to_cache}
\alias{save_to_cache}
\title{Save file as an RDS to indycar cache}
\usage{
save_to_cache(df, filename)
}
\arguments{
\item{df}{a tibble}

\item{filename}{the name of the file (should match the original name of the file read in)}
}
\value{
None
}
\description{
Save file as an RDS to indycar cache
}
