% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_tables.R
\name{table_pit_strategy_lap_times}
\alias{table_pit_strategy_lap_times}
\title{Green Flag Lap Times by number of stops}
\usage{
table_pit_strategy_lap_times(df, drivers)
}
\arguments{
\item{df}{(tibble): a laps df}

\item{drivers}{(list): a list of drivers. Defaults to all drivers.}
}
\value{
a gt object
}
\description{
Green Flag Lap Times by number of stops
}
