% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_tables.R
\docType{data}
\name{team_colors}
\alias{team_colors}
\title{List of teams and colors}
\format{
An object of class \code{character} of length 10.
}
\usage{
team_colors
}
\description{
Team names and primary car colors
}
\keyword{datasets}
