% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_tables.R
\docType{data}
\name{team_manufacturers}
\alias{team_manufacturers}
\title{List of teams and enging manufacturers}
\format{
An object of class \code{character} of length 10.
}
\usage{
team_manufacturers
}
\description{
Team name and engine manufacturer name.
}
\keyword{datasets}
