% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_timing.R
\name{filter_yellow}
\alias{filter_yellow}
\title{Filter Out Yellow Flag Laps}
\usage{
filter_yellow(df)
}
\arguments{
\item{df}{a laps df}
}
\value{
tibble
}
\description{
Filters out yellow flag laps.
}
\examples{
no_yellows <- laps \%>\% filter_yellow()
}
