% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{laps}
\alias{laps}
\title{Laps Data}
\format{
\subsection{\code{laps}}{

A data frame with 2262 rows and 23 columns:
}
}
\source{
\url{https://racetools.com/logfiles/IndyCar/2023/Gallagher\%20Grand\%20Prix(Indianapolis\%20Motor\%20Speedway\%20RC)-Race_(R.I)_2023-08-12.zip}
}
\usage{
laps
}
\description{
A dataframe of race laps from 2023 Indianapolis Road Course GP.
}
\keyword{datasets}
